requirements.txt
flask
flask-cors
python-dotenv
requests
solders
openai
